+++ Headers/EBO.h
@@ -12,7 +12,7 @@ class EBO
 {
     public:
         GLuint ID;
-        EBO(GLfloat* indices,GLsizeiptr size);
+        EBO(GLuint* indices,GLsizeiptr size);
 
         void Bind();
         void Unbind();
